hasHtmlTableStyle {htmlTable}	R Documentation

Check if object has a style set to it

Description

If the attribute htmlTable.style is set it will check if the style_name exists and return a logical.

Usage

hasHtmlTableStyle(x, style_name)

Arguments

x
The object intended for htmlTable().

style_name
A string that contains the style name.

Value

logical TRUE if the attribute and style is not NULL

See Also

Other htmlTableStyle: addHtmlTableStyle()

Examples

library(magrittr)

mx <- matrix(1:4, ncol = 2)
colnames(mx) <- LETTERS[1:2]
mx %>%
  addHtmlTableStyle(align = "l|r") %>%
  hasHtmlTableStyle("align")

[Package htmlTable version 2.4.0 Index]