htmlTableWidget-shiny {htmlTable}	R Documentation

Shiny bindings for htmlTableWidget

Description

Output and render functions for using htmlTableWidget within Shiny applications and interactive Rmd documents.

Usage

htmlTableWidgetOutput(outputId, width = "100%", height = "400px")

renderHtmlTableWidget(expr, env = parent.frame(), quoted = FALSE)

Arguments

outputId
output variable to read from

width, height
Must be a valid CSS unit (like '100%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended.

expr
An expression that generates a htmlTableWidget

env
The environment in which to evaluate expr.

quoted
Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.

Examples

## Not run: 
# In the UI:
htmlTableWidgetOutput("mywidget")
# In the server:
renderHtmlTableWidget({htmlTableWidget(iris)})

## End(Not run)

[Package htmlTable version 1.13.3 Index]