innerJoinByCommonCols {htmlTable}	R Documentation

A simple function for joining two tables by their intersected columns

Description

A simple function for joining two tables by their intersected columns

Usage

innerJoinByCommonCols(x, y)

Arguments

x
data.frame

y
data.frame

Value

data.frame

[Package htmlTable version 2.4.0 Index]