prAddCells {htmlTable}	R Documentation

Add a cell

Description

Adds a row of cells <td>val</td><td>...</td> to a table string for htmlTable()

Usage

prAddCells(
  rowcells,
  cellcode,
  style_list,
  style,
  prepped_cell_css,
  cgroup_spacer_cells,
  has_rn_col,
  offset = 1,
  style_list_align_key = "align"
)

Arguments

rowcells
The cells with the values that are to be added

cellcode
Type of cell, can either be th or td

style_list
The style_list

style
The cell style

cgroup_spacer_cells
The number of cells that occur between columns due to the cgroup arguments.

has_rn_col
Due to the alignment issue we need to keep track of if there has already been printed a rowname column or not and therefore we have this has_rn_col that is either 0 or 1.

offset
For rgroup rows there may be an offset != 1

Value

string Returns the string with the new cell elements

See Also

Other hidden helper functions for htmlTable: prAddEmptySpacerCell(), prAddSemicolon2StrEnd(), prEscapeHtml(), prGetCgroupHeader(), prGetRowlabelPos(), prGetStyle(), prPrepInputMatrixDimensions(), prPrepareAlign(), prPrepareCgroup(), prTblNo()

[Package htmlTable version 2.4.0 Index]