prAddSemicolon2StrEnd {htmlTable}	R Documentation

Add a ; at the end

Description

The CSS expects a semicolon at the end of each argument this function just adds a semicolong if none is given and remove multiple semicolon if such exist

Usage

prAddSemicolon2StrEnd(my_str)

Arguments

my_str
The string that is to be processed

Value

string

See Also

Other hidden helper functions for htmlTable: prAddCells(), prEscapeHtml(), prGetCgroupHeader(), prGetRowlabelPos(), prGetStyle(), prPrepareAlign(), prPrepareCgroup(), prTblNo()

[Package htmlTable version 1.13.3 Index]