prBindDataListIntoColumns {htmlTable}	R Documentation

Merge columns into a tibble

Description

Almost the same as tibble::tibble() but it solves the issue with some of the arguments being columns and some just being vectors.

Usage

prBindDataListIntoColumns(dataList)

Arguments

dataList
list with the columns/data.frames

Value

data.frame object

[Package htmlTable version 2.4.0 Index]