prExtractElementsAndConvertToTbl {htmlTable}	R Documentation

Extract the elements and generate a table with unique elements

Description

Extract the elements and generate a table with unique elements

Usage

prExtractElementsAndConvertToTbl(x, elements)

Arguments

x
list with columns to be joined

elements
char vector with the elements to select

[Package htmlTable version 2.4.0 Index]