prGetAlign {htmlTable}	R Documentation

Gets alignment

Description

Gets alignment

Usage

prGetAlign(align, index)

Arguments

align
A character strings specifying column alignments, defaulting to paste(rep('c',ncol(x)),collapse='') to center. Valid alignments are l = left, c = center and r = right. You can also specify align='c|c' and other LaTeX tabular formatting. If you want to set the alignment of the rownames this string needst to be ncol(x) + 1, otherwise it automatically pads the string with a left alignment for the rownames.

index
The index of the align parameter of interest

[Package htmlTable version 1.13.3 Index]