prGetRgroupLine {htmlTable}	R Documentation

Gets the number of rgroup HTML line

Description

Gets the number of rgroup HTML line

Usage

prGetRgroupLine(
  x,
  total_columns = NULL,
  rgroup = NULL,
  rgroup_iterator = NULL,
  cspan = NULL,
  rnames = NULL,
  style = NULL,
  cgroup_spacer_cells = NULL,
  style_list = NULL,
  prepped_row_css = NULL
)

Arguments

x
The matrix/data.frame with the data. For the print and knit_print it takes a string of the class htmlTable as x argument.

total_columns
The total number of columns including the rowlabel and the spacer cells

rgroup
A vector of character strings containing headings for row groups. n.rgroup must be present when rgroup is given. See detailed description in section below.

rgroup_iterator
An integer indicating the rgroup

cspan
The column span of the current rgroup

rnames
Default row names are generated from rownames(x). If you provide FALSE then it will skip the row names. Note: For data.frames if you do rownames(my_dataframe) <- NULL it still has row names. Thus you need to use FALSE if you want to supress row names for data.frames.

style
The css style corresponding to the rgroup css style that includes the color specific for the rgroup, i.e. col.rgroup.

cgroup_spacer_cells
The vector indicating the position of the cgroup spacer cells

prepped_row_css
The css.cell information for this particular row.

[Package htmlTable version 2.4.0 Index]