prGetRowlabelPos {htmlTable}	R Documentation

Gets the rowlabel position

Description

Gets the rowlabel position

Usage

prGetRowlabelPos(cgroup, pos.rowlabel, header)

Arguments

cgroup
A vector, matrix or list of character strings defining major column header. The default is to have none. These elements are also known as column spanners. If you want a column not to have a spanner then put that column as "". If you pass cgroup and n.crgroup as matrices you can have column spanners for several rows. See cgroup section below for details.

pos.rowlabel
Where the rowlabel should be positioned. This value can be "top", "bottom", "header", or a integer between 1 and nrow(cgroup) + 1. The options "bottom" and "header" are the same, where the row label is presented at the same level as the header.

header
A vector of character strings specifying column header, defaulting to colnames(x)

Value

integer Returns the position within the header rows to print the rowlabel argument

See Also

Other hidden helper functions for htmlTable: prAddCells(), prAddSemicolon2StrEnd(), prEscapeHtml(), prGetCgroupHeader(), prGetStyle(), prPrepareAlign(), prPrepareCgroup(), prTblNo()

[Package htmlTable version 1.13.3 Index]