prGetScriptString {htmlTable}	R Documentation

Gets a string with all the scripts merged into one script tag

Description

Each element has it's own script tags in otherwise an error will cause all the scripts to fail.

Usage

prGetScriptString(x)

Arguments

x
An interactiveTable

Value

string

[Package htmlTable version 1.13.3 Index]