prGetStyle {htmlTable}	R Documentation

Gets the CSS style element

Description

A function for checking, merging, and more with a variety of different style formats.

Usage

prGetStyle(...)

Arguments

...
Styles can be provided as vector, named vector, or string. If you provide a name, e.g. background: blue, align="center", the function will convert the align into proper align: center.

Value

string Returns the codes merged into one string with correct CSS ; and : structure.

See Also

Other hidden helper functions for htmlTable: prAddCells(), prAddEmptySpacerCell(), prAddSemicolon2StrEnd(), prEscapeHtml(), prGetCgroupHeader(), prGetRowlabelPos(), prPrepInputMatrixDimensions(), prPrepareAlign(), prPrepareCgroup(), prTblNo()

[Package htmlTable version 2.4.2 Index]