prIsNotebook {htmlTable}	R Documentation

Detects if the call is made from within an RStudio Rmd file or a file with the html_notebook output set.

Description

Detects if the call is made from within an RStudio Rmd file or a file with the html_notebook output set.

Usage

prIsNotebook()

[Package htmlTable version 1.13.3 Index]