prMergeClr {htmlTable}	R Documentation

Merges multiple colors

Description

Uses the colorRampPalette for merging colors. Note: When merging more than 2 colors the order in the color presentation matters. Each color is merged with its neigbors before merging with next. If there is an uneven number of colors the middle color is mixed with both left and right side.

Usage

prMergeClr(clrs)

Arguments

clrs
The colors

Value

character A hexadecimal color

[Package htmlTable version 1.13.3 Index]