prPrepInputMatrixDimensions {htmlTable}	R Documentation

Makes sure the input is correct

Description

Checks and converts dimensions into something the htmlTable() is comfortable with.

Usage

prPrepInputMatrixDimensions(x, header = NULL)

Arguments

x
The matrix/data.frame with the data. For the print and knit_print it takes a string of the class htmlTable as x argument.

header
A vector of character strings specifying column header, defaulting to colnames(x)

See Also

Other hidden helper functions for htmlTable: prAddCells(), prAddEmptySpacerCell(), prAddSemicolon2StrEnd(), prEscapeHtml(), prGetCgroupHeader(), prGetRowlabelPos(), prGetStyle(), prPrepareAlign(), prPrepareCgroup(), prTblNo()

[Package htmlTable version 2.4.2 Index]