prPrepareAlign {htmlTable}	R Documentation

Prepares the align to match the columns

Description

The alignment may be tricky and this function therefore simplifies this process by extending/shortening the alignment to match the correct number of columns.

Usage

prPrepareAlign(align, x, rnames, default_rn = "l")

Arguments

x
The matrix/data.frame with the data. For the print and knit_print it takes a string of the class htmlTable as x argument.

rnames
Default row names are generated from rownames(x). If you provide FALSE then it will skip the row names. Note: For data.frames if you do rownames(my_dataframe) <- NULL it still has row names. Thus you need to use FALSE if you want to supress row names for data.frames.

default_rn
The default rowname alignment. This is an option as the header uses the same function and there may be differences in how the alignments should be implemented.

See Also

Other hidden helper functions for htmlTable: prAddCells(), prAddEmptySpacerCell(), prAddSemicolon2StrEnd(), prEscapeHtml(), prGetCgroupHeader(), prGetRowlabelPos(), prGetStyle(), prPrepInputMatrixDimensions(), prPrepareCgroup(), prTblNo()

[Package htmlTable version 2.4.2 Index]