prPrepareCgroup {htmlTable}	R Documentation

Prepares the cgroup argument

Description

Due to the complicated structure of multilevel cgroups there some preparation for the cgroup options is required.

Usage

prPrepareCgroup(x, cgroup, n.cgroup, align.cgroup, css.cgroup)

Arguments

x
The matrix/data.frame with the data. For the print and knit_print it takes a string of the class htmlTable as x argument.

cgroup
A vector, matrix or list of character strings defining major column header. The default is to have none. These elements are also known as column spanners. If you want a column not to have a spanner then put that column as "". If you pass cgroup and n.crgroup as matrices you can have column spanners for several rows. See cgroup section below for details.

n.cgroup
An integer vector, matrix or list containing the number of columns for which each element in cgroup is a heading. For example, specify cgroup=c("Major_1","Major_2"), n.cgroup=c(3,3) if "Major_1" is to span columns 1-3 and "Major_2" is to span columns 4-6. rowlabel does not count in the column numbers. You can omit n.cgroup if all groups have the same number of columns. If the n.cgroup is one less than the number of columns in the matrix/data.frame then it automatically adds those.

align.cgroup
The justification of the cgroups

css.cgroup
The same as css.class but for cgroup formatting.

Value

list(cgroup, n.cgroup, align.cgroup, cgroup_spacer_cells)

See Also

Other hidden helper functions for htmlTable: prAddCells(), prAddSemicolon2StrEnd(), prEscapeHtml(), prGetCgroupHeader(), prGetRowlabelPos(), prGetStyle(), prPrepareAlign(), prTblNo()

[Package htmlTable version 1.13.3 Index]