prPrepareCss {htmlTable}	R Documentation

Prepares the cell style

Description

Prepares the cell style

Usage

prPrepareCss(x, css, rnames, header, name = deparse(substitute(css)))

Arguments

x
The matrix/data.frame with the data. For the print and knit_print it takes a string of the class htmlTable as x argument.

css
The CSS styles that are to be converted into a matrix.

rnames
Default rownames are generated from rownames(x). If you provide FALSE then it will skip the rownames. Note: For data.frames if you do rownames(my_dataframe) <- NULL it still has rownames. Thus you need to use FALSE if you want to surpress rownames for data.frames.

header
A vector of character strings specifying column header, defaulting to colnames(x)

name
The name of the CSS style that is prepared

Value

matrix

[Package htmlTable version 1.13.3 Index]