prSkipRownames {htmlTable}	R Documentation

Returns if rownames should be printed for the htmlTable

Description

Returns if rownames should be printed for the htmlTable

Usage

prSkipRownames(rnames)

Arguments

rnames
Default rownames are generated from rownames(x). If you provide FALSE then it will skip the rownames. Note: For data.frames if you do rownames(my_dataframe) <- NULL it still has rownames. Thus you need to use FALSE if you want to surpress rownames for data.frames.

[Package htmlTable version 1.13.3 Index]