prTblNo {htmlTable}	R Documentation

Gets the table counter string

Description

Returns the string used for htmlTable to number the different tables. Uses options table_counter, table_counter_str, and table_counter_roman to produce the final string. You can set each option by simply calling options().

Usage

prTblNo(caption = NULL)

Arguments

caption
The caption if any

Value

string Returns a string formatted according to the table_counter_str and table_counter_roman. The number is decided by the table_counter variable

See Also

Other hidden helper functions for htmlTable: prAddCells(), prAddEmptySpacerCell(), prAddSemicolon2StrEnd(), prEscapeHtml(), prGetCgroupHeader(), prGetRowlabelPos(), prGetStyle(), prPrepInputMatrixDimensions(), prPrepareAlign(), prPrepareCgroup()

[Package htmlTable version 2.4.2 Index]