tblNoLast {htmlTable}	R Documentation

Gets the last table number

Description

The function relies on options("table_counter") in order to keep track of the last number.

Usage

tblNoLast(roman = getOption("table_counter_roman", FALSE))

Arguments

roman
Whether or not to use roman numbers instead of arabic. Can also be set through options(table_caption_no_roman = TRUE)

See Also

Other table functions: htmlTable, tblNoNext()

Examples

org_opts <- options(table_counter=1)
tblNoLast()
options(org_opts)

[Package htmlTable version 1.13.3 Index]