txtPval {htmlTable}	R Documentation

Formats the p-values

Description

Gets formatted p-values. For instance you often want 0.1234 to be 0.12 while also having two values up until a limit, i.e. 0.01234 should be 0.012 while 0.001234 should be 0.001. Furthermore you want to have < 0.001 as it becomes ridiculous to report anything below that value.

Usage

txtPval(pvalues, lim.2dec = 10^-2, lim.sig = 10^-4, html = TRUE, ...)

Arguments

pvalues
The p-values

lim.2dec
The limit for showing two decimals. E.g. the p-value may be 0.056 and we may want to keep the two decimals in order to emphasize the proximity to the all-mighty 0.05 p-value and set this to 10^-2. This allows that a value of 0.0056 is rounded to 0.006 and this makes intuitive sense as the 0.0056 level as this is well below the 0.05 value and thus not as interesting to know the exact proximity to 0.05. Disclaimer: The 0.05-limit is really silly and debated, unfortunately it remains a standard and this package tries to adapt to the current standards in order to limit publication associated issues.

lim.sig
The significance limit for the less than sign, i.e. the '<'

html
If the less than sign should be < or &lt; as needed for html output.

...
Currently only used for generating warnings of deprecated call parameters.

Value

vector

See Also

Other text formatters: txtMergeLines(), txtRound()

Examples

txtPval(c(0.10234,0.010234, 0.0010234, 0.000010234))

[Package htmlTable version 1.13.3 Index]