prAddEmptySpacerCell {htmlTable}	R Documentation

Add an empty cell

Description

Depending on the spacer.celltype set in addHtmlTableStyle() we will use different spacer cells.

Usage

prAddEmptySpacerCell(
  x,
  style_list,
  cell_style,
  align_style,
  cell_tag = c("td", "th"),
  colspan = 1
)

Arguments

x
The matrix/data.frame with the data. For the print and knit_print it takes a string of the class htmlTable as x argument.

cell_style
The style of the current cell that should be applied to all cells

align_style
The style from prGetAlign()

cell_tag
What HTML tag to use

colspan
The number of rows each tag should span

Value

string

See Also

Other hidden helper functions for htmlTable: prAddCells(), prAddSemicolon2StrEnd(), prEscapeHtml(), prGetCgroupHeader(), prGetRowlabelPos(), prGetStyle(), prPrepInputMatrixDimensions(), prPrepareAlign(), prPrepareCgroup(), prTblNo()

[Package htmlTable version 2.4.0 Index]