prGetRgroupLine {htmlTable}	R Documentation

Gets the number of rgroup htmlLine

Description

Gets the number of rgroup htmlLine

Usage

prGetRgroupLine(
  x,
  total_columns,
  rgroup,
  rgroup_iterator,
  cspan,
  rnames,
  align,
  style,
  cgroup_spacer_cells,
  col.columns,
  css.row,
  padding.tspanner
)

Arguments

x
The matrix/data.frame with the data. For the print and knit_print it takes a string of the class htmlTable as x argument.

total_columns
The total number of columns including the rowlabel and the spacer cells

rgroup
A vector of character strings containing headings for row groups. n.rgroup must be present when rgroup is given. See detailed description in section below.

rgroup_iterator
An integer indicating the rgroup

cspan
The column span of the current rgroup

rnames
Default rownames are generated from rownames(x). If you provide FALSE then it will skip the rownames. Note: For data.frames if you do rownames(my_dataframe) <- NULL it still has rownames. Thus you need to use FALSE if you want to surpress rownames for data.frames.

align
A character strings specifying column alignments, defaulting to paste(rep('c',ncol(x)),collapse='') to center. Valid alignments are l = left, c = center and r = right. You can also specify align='c|c' and other LaTeX tabular formatting. If you want to set the alignment of the rownames this string needst to be ncol(x) + 1, otherwise it automatically pads the string with a left alignment for the rownames.

style
The css style corresponding to the rgroup css style that includes the color specific for the rgroup, i.e. col.rgroup.

cgroup_spacer_cells
The vector indicating the position of the cgroup spacer cells

col.columns
Alternating colors for each column.

css.row
The css.cell information for this particular row.

padding.tspanner
The tspanner padding

[Package htmlTable version 1.13.3 Index]