prPrepareAlign {htmlTable}	R Documentation

Prepares the align to match the columns

Description

The alignment may be tricky and this function therefore simplifies this process by extending/shortening the alignment to match the correct number of columns.

Usage

prPrepareAlign(align, x, rnames, default_rn = "l")

Arguments

align
A character strings specifying column alignments, defaulting to paste(rep('c',ncol(x)),collapse='') to center. Valid alignments are l = left, c = center and r = right. You can also specify align='c|c' and other LaTeX tabular formatting. If you want to set the alignment of the rownames this string needst to be ncol(x) + 1, otherwise it automatically pads the string with a left alignment for the rownames.

x
The matrix/data.frame with the data. For the print and knit_print it takes a string of the class htmlTable as x argument.

rnames
Default rownames are generated from rownames(x). If you provide FALSE then it will skip the rownames. Note: For data.frames if you do rownames(my_dataframe) <- NULL it still has rownames. Thus you need to use FALSE if you want to surpress rownames for data.frames.

default_rn
The default rowname alignment. This is an option as the header uses the same function and there may be differences in how the alignments should be implemented.

See Also

Other hidden helper functions for htmlTable: prAddCells(), prAddSemicolon2StrEnd(), prEscapeHtml(), prGetCgroupHeader(), prGetRowlabelPos(), prGetStyle(), prPrepareCgroup(), prTblNo()

[Package htmlTable version 1.13.3 Index]